+++ src/HomeAssistant/Sensor.ts
@@ -4,7 +4,7 @@ import { EntityConfig } from './base/Entity';
 import { StatefulEntity } from './base/StatefulEntity';
 
 export class Sensor<T> extends StatefulEntity<T> {
-  constructor(mqtt: IMQTTConnection, deviceData: IDeviceData, entityConfig: EntityConfig) {
-    super(mqtt, deviceData, entityConfig, 'sensor');
+  constructor(mqtt: IMQTTConnection, deviceData: IDeviceData, entityConfig: EntityConfig, retainState: boolean = false) {
+    super(mqtt, deviceData, entityConfig, 'sensor', retainState);
   }
 }