+++ src/MQTT/IMQTTConnection.ts
@@ -2,6 +2,6 @@ export interface IMQTTConnection {
   unsubscribe(topic: string): void;
   on(event: string, listener: (this: IMQTTConnection, message: string) => void): this;
   off(event: string, listener: (this: IMQTTConnection, message: string) => void): this;
-  publish(topic: string, message: any): void;
+  publish(topic: string, message: any, retain?: boolean): void;
   subscribe(topic: string): void;
 }